# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PadelHub API - Sistema de Gestión de Partidos",
    "version": "1.4.0",
    "description": "Documentación técnica oficial del backend PadelHub."
  },
  "servers": [
    {
      "url": "https://padelhub-backend-phi.vercel.app/api",
      "description": "Servidor de producción en Vercel"
    }
  ],
...
  "paths": {
...
    "/leaderboard": {
      "get": {
        "tags": ["8. Leaderboard"],
        "summary": "Ranking de jugadores por MMR",
        "description": "Devuelve el ranking ordenado por MMR descendente. Sin parámetros retorna el ranking nacional. Usar `zone` para el ranking por zona.",
        "parameters": [
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Filtra el ranking por zona (ej: 'Santiago', 'Concepción'). Sin este parámetro se retorna el ranking nacional.",
            "schema": { "type": "string", "example": "Santiago" }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "description": "Filtra por nivel de jugador",
            "schema": {
              "type": "string",
              "enum": ["primera", "segunda", "tercera", "cuarta", "quinta", "sexta", "septima_mas"],
              "example": "segunda"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Máximo de resultados a retornar (default: 50, máx: 100)",
            "schema": { "type": "integer", "example": 50, "default": 50, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking de jugadores.",
            "content": {
              "application/json": {
                "example": {
                  "scope": "zone",
                  "zone": "Santiago",
                  "level": null,
                  "total": 23,
                  "players": [
                    {
                      "rank": 1,
                      "id": "9d39cac8-e836-41c5-847b-bfb6459a9939",
                      "name": "Juan Pérez",
                      "photo_url": "https://res.cloudinary.com/padelhub/image/upload/padelhub_avatars/abc123.jpg",
                      "level": "primera",
                      "zone": "Santiago",
                      "mmr": 1340,
                      "matches_played": 12
                    }
                  ]
                }
              }
            }
          },
          "500": { "description": "Error interno al obtener el leaderboard." }
        }
      }
    }
  }
}